Move buildTree from FileExplorer to commitUtils
