--- explorer/src/FileExplorer.test.js
@@ -1,14 +1,0 @@
-import {buildTree} from './FileExplorer';
-
-describe('buildTree', () => {
-  it('handles empty tree', () => {
-    expect(buildTree([])).toEqual({});  
-  });
-
-  it('handles trees', () => {
-    const names = ['foo', 'bar/zod', 'bar/zoink'];
-    const expected = {'foo': {}, 'bar': {'zod': {}, 'zoink': {}}};
-    expect(buildTree(names)).toEqual(expected);
-  });
-});
-
